Updated artifact publish to take place on a remote slave if one is used
